Migrate to react-navigation
