--- .eslintignore
@@ -1,2 +1,0 @@
-/.expo
-node_modules
