--- .github/pull_request_template.md
@@ -1,5 +1,0 @@
-## Changes
-
--
-
-## Screenshots
