+++ babel.config.js
@@ -2,11 +2,6 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: ["babel-preset-expo"],
-    plugins: [
-      "@babel/plugin-proposal-export-namespace-from",
-      require.resolve("expo-router/babel"),
-      "nativewind/babel",
-      "react-native-reanimated/plugin",
-    ],
+    plugins: ["nativewind/babel", "react-native-reanimated/plugin"],
   };
 };
