--- metro.config.js
@@ -1,9 +1,0 @@
-// metro.confgi.js
-const { getDefaultConfig } = require("expo/metro-config");
-
-// eslint-disable-next-line no-undef
-const config = getDefaultConfig(__dirname, {
-  isCSSEnabled: true,
-});
-
-module.exports = config;
