+++ package.json
@@ -1,71 +1,51 @@
 {
   "name": "crema-mobile",
-  "version": "0.0.3",
-  "main": "expo-router/entry",
+  "version": "0.0.4",
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web",
     "format": "prettier --write .",
-    "lint": "npm run format && eslint . --fix"
+    "lint": "npm run format && tsc && eslint . --fix"
   },
   "dependencies": {
     "@react-native-community/slider": "4.4.2",
-    "@react-native-picker/picker": "2.4.10",
+    "@react-navigation/bottom-tabs": "^6.5.9",
     "@react-navigation/native": "^6.1.7",
+    "@react-navigation/native-stack": "^6.9.13",
     "@tanstack/react-query": "^4.32.0",
     "@trpc/client": "^10.36.0",
     "@trpc/react-query": "^10.36.0",
     "@trpc/server": "^10.36.0",
+    "@types/react": "~18.2.14",
     "date-fns": "^2.30.0",
-    "expo": "^49.0.7",
-    "expo-blur": "~12.4.1",
-    "expo-constants": "~14.4.2",
-    "expo-dev-client": "~2.4.8",
+    "expo": "~49.0.11",
     "expo-haptics": "~12.4.0",
-    "expo-image": "~1.3.2",
     "expo-image-picker": "~14.3.2",
-    "expo-linking": "~5.0.2",
-    "expo-router": "2.0.0",
     "expo-secure-store": "~12.3.1",
-    "expo-splash-screen": "~0.20.5",
     "expo-status-bar": "~1.6.0",
-    "expo-system-ui": "~2.4.0",
     "formik": "^2.4.3",
     "lucide-react-native": "^0.265.0",
     "nativewind": "^2.0.11",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.72.4",
-    "react-native-figma-squircle": "^0.3.4",
-    "react-native-gesture-handler": "~2.12.0",
+    "react-native": "0.72.5",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
-    "react-native-svg": "^13.13.0",
-    "react-native-web": "~0.19.6",
+    "react-native-svg": "13.9.0",
     "react-timer-hook": "^3.0.7",
     "typescript": "^5.1.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.3",
-    "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
-    "@types/react": "~18.2.14",
+    "@babel/core": "^7.20.0",
     "eslint": "^8.46.0",
     "eslint-config-universe": "^11.3.0",
     "eslint-plugin-prettier": "^5.0.0",
     "prettier": "^3.0.1",
     "tailwindcss": "3.3.2"
-  },
-  "resolutions": {
-    "metro": "^0.73.7",
-    "metro-resolver": "^0.73.7"
-  },
-  "overrides": {
-    "metro": "^0.73.7",
-    "metro-resolver": "^0.73.7"
   },
   "private": true
 }
