+++ screens/login.tsx
@@ -1,4 +1,4 @@
-import { Link } from "expo-router";
+import { useNavigation } from "@react-navigation/native";
 import { Formik } from "formik";
 import { ArrowRight } from "lucide-react-native";
 import { useContext } from "react";
@@ -8,30 +8,32 @@ import {
   Platform,
   ScrollView,
   Text,
+  TouchableOpacity,
   View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import Button from "../../components/button";
-import Input from "../../components/input";
-import AuthContext from "../../lib/context/auth";
-import { trpc } from "../../lib/trpc";
+import Button from "../components/button";
+import Input from "../components/input";
+import AuthContext from "../lib/context/auth";
+import { trpc } from "../lib/trpc";
+import { LoginNavigationProp } from "../types/navigation";
 
-export default function Signup() {
+export default function Login() {
   const insets = useSafeAreaInsets();
+  const navigation = useNavigation<LoginNavigationProp>();
 
   // Global state
   const { setToken } = useContext(AuthContext);
 
   // Server state
-  const { mutateAsync, isLoading } = trpc.auth.signup.useMutation({
+  const { mutateAsync, isLoading } = trpc.auth.login.useMutation({
     onSuccess: (data) => {
       setToken(data.token);
     },
     onError: () => {
-      Alert.alert("Failed to sign up");
+      Alert.alert("Failed to log in");
     },
-    retry: false,
   });
 
   return (
@@ -43,10 +45,8 @@ export default function Signup() {
       <ScrollView>
         <Formik
           initialValues={{
-            name: "",
             email: "",
             password: "",
-            confirmPassword: "",
           }}
           onSubmit={(values) => {
             mutateAsync(values);
@@ -55,7 +55,7 @@ export default function Signup() {
           {({ handleChange, handleSubmit, values }) => (
             <View className="p-5 pt-16">
               <Text className="mt-8 text-4xl font-semibold dark:text-white">
-                Sign up to Crema
+                Welcome to Crema
               </Text>
 
               <View className="mt-6">
@@ -65,51 +65,41 @@ export default function Signup() {
                   clearButtonMode="while-editing"
                   keyboardType="email-address"
                   autoCapitalize="none"
+                  autoComplete="email"
                   autoCorrect={false}
                   value={values.email}
                   label="Email"
                 />
 
                 <Input
-                  onChangeText={handleChange("name")}
-                  clearButtonMode="while-editing"
-                  placeholder="Johanna Doe"
-                  autoCapitalize="words"
-                  value={values.name}
-                  className="mt-7"
-                  label="Name"
-                />
-
-                <Input
                   onChangeText={handleChange("password")}
                   clearButtonMode="while-editing"
                   placeholder="••••••••••••••"
-                  autoComplete="new-password"
                   value={values.password}
+                  autoComplete="password"
                   autoCapitalize="none"
+                  autoCorrect={false}
                   label="Password"
-                  className="mt-7"
-                  secureTextEntry
-                />
-
-                <Input
-                  onChangeText={handleChange("confirmPassword")}
-                  clearButtonMode="while-editing"
-                  placeholder="••••••••••••••"
-                  value={values.confirmPassword}
-                  label="Confirm password"
-                  autoCapitalize="none"
                   className="mt-7"
                   secureTextEntry
                 />
               </View>
 
-              <Text className="mt-8 text-base dark:text-white">
-                Already have a profile?{" "}
-                <Link href="/" className="text-indigo-400 dark:text-indigo-300">
-                  Log in here
-                </Link>
-              </Text>
+              <View className="flex-row items-center gap-2 flex-wrap mt-8">
+                <Text className="text-base dark:text-white">
+                  Don't have a profile yet?
+                </Text>
+
+                <TouchableOpacity
+                  onPress={() => {
+                    navigation.navigate("Signup");
+                  }}
+                >
+                  <Text className="text-base text-indigo-400 dark:text-indigo-300">
+                    Sign up here
+                  </Text>
+                </TouchableOpacity>
+              </View>
 
               <Button
                 onPress={() => handleSubmit()}
@@ -117,7 +107,7 @@ export default function Signup() {
                 className="mt-10"
                 icon={ArrowRight}
               >
-                Sign up
+                Log in
               </Button>
             </View>
           )}
