+++ screens/profile.tsx
@@ -1,9 +1,9 @@
 import { LogOutIcon } from "lucide-react-native";
 import { useContext } from "react";
 
-import Button from "../../components/button";
-import Screen from "../../components/screen";
-import AuthContext from "../../lib/context/auth";
+import Button from "../components/button";
+import Screen from "../components/screen";
+import AuthContext from "../lib/context/auth";
 
 export default function Profile() {
   const { setToken } = useContext(AuthContext);
