+++ screens/signup.tsx
@@ -1,4 +1,4 @@
-import { Link } from "expo-router";
+import { useNavigation } from "@react-navigation/native";
 import { Formik } from "formik";
 import { ArrowRight } from "lucide-react-native";
 import { useContext } from "react";
@@ -8,29 +8,33 @@ import {
   Platform,
   ScrollView,
   Text,
+  TouchableOpacity,
   View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import Button from "../../components/button";
-import Input from "../../components/input";
-import AuthContext from "../../lib/context/auth";
-import { trpc } from "../../lib/trpc";
+import Button from "../components/button";
+import Input from "../components/input";
+import AuthContext from "../lib/context/auth";
+import { trpc } from "../lib/trpc";
+import { SignupNavigationProp } from "../types/navigation";
 
-export default function Login() {
+export default function Signup() {
   const insets = useSafeAreaInsets();
+  const navigation = useNavigation<SignupNavigationProp>();
 
   // Global state
   const { setToken } = useContext(AuthContext);
 
   // Server state
-  const { mutateAsync, isLoading } = trpc.auth.login.useMutation({
+  const { mutateAsync, isLoading } = trpc.auth.signup.useMutation({
     onSuccess: (data) => {
       setToken(data.token);
     },
     onError: () => {
-      Alert.alert("Failed to log in");
+      Alert.alert("Failed to sign up");
     },
+    retry: false,
   });
 
   return (
@@ -42,8 +46,10 @@ export default function Login() {
       <ScrollView>
         <Formik
           initialValues={{
+            name: "",
             email: "",
             password: "",
+            confirmPassword: "",
           }}
           onSubmit={(values) => {
             mutateAsync(values);
@@ -52,7 +58,7 @@ export default function Login() {
           {({ handleChange, handleSubmit, values }) => (
             <View className="p-5 pt-16">
               <Text className="mt-8 text-4xl font-semibold dark:text-white">
-                Welcome to Crema
+                Sign up to Crema
               </Text>
 
               <View className="mt-6">
@@ -62,35 +68,60 @@ export default function Login() {
                   clearButtonMode="while-editing"
                   keyboardType="email-address"
                   autoCapitalize="none"
-                  autoComplete="email"
                   autoCorrect={false}
                   value={values.email}
                   label="Email"
                 />
 
                 <Input
+                  onChangeText={handleChange("name")}
+                  clearButtonMode="while-editing"
+                  placeholder="Johanna Doe"
+                  autoCapitalize="words"
+                  value={values.name}
+                  className="mt-7"
+                  label="Name"
+                />
+
+                <Input
                   onChangeText={handleChange("password")}
                   clearButtonMode="while-editing"
                   placeholder="••••••••••••••"
+                  autoComplete="new-password"
                   value={values.password}
-                  autoComplete="password"
                   autoCapitalize="none"
-                  autoCorrect={false}
                   label="Password"
+                  className="mt-7"
+                  secureTextEntry
+                />
+
+                <Input
+                  onChangeText={handleChange("confirmPassword")}
+                  clearButtonMode="while-editing"
+                  placeholder="••••••••••••••"
+                  value={values.confirmPassword}
+                  label="Confirm password"
+                  autoCapitalize="none"
                   className="mt-7"
                   secureTextEntry
                 />
               </View>
 
-              <Text className="mt-8 text-base dark:text-white">
-                Don't have a profile yet?{" "}
-                <Link
-                  href="/signup"
-                  className="text-indigo-400 dark:text-indigo-300"
+              <View className="flex-row items-center gap-2 flex-wrap mt-8">
+                <Text className="mt-8 text-base dark:text-white">
+                  Already have a profile?
+                </Text>
+
+                <TouchableOpacity
+                  onPress={() => {
+                    navigation.navigate("Login");
+                  }}
                 >
-                  Sign up here
-                </Link>
-              </Text>
+                  <Text className="text-base text-indigo-400 dark:text-indigo-300">
+                    Log in here
+                  </Text>
+                </TouchableOpacity>
+              </View>
 
               <Button
                 onPress={() => handleSubmit()}
@@ -98,7 +129,7 @@ export default function Login() {
                 className="mt-10"
                 icon={ArrowRight}
               >
-                Log in
+                Sign up
               </Button>
             </View>
           )}
